pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "infraguard"
version = "0.1.0"
description = "Infrastructure guardrails for teams that ship fast — Terraform risk scoring, AWS tag auditing, and IAM policy analysis in one CLI."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Bertrand Mbanwi", email = "bertrandmbanwi@gmail.com" },
]
keywords = [
    "terraform",
    "aws",
    "iam",
    "security",
    "devops",
    "cloud",
    "infrastructure",
    "compliance",
    "blast-radius",
    "least-privilege",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",

    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
aws = ["boto3>=1.34.0"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
]
all = ["infraguard[aws,dev]"]

[project.urls]
Homepage = "https://github.com/bertrandmbanwi/infraguard"
Repository = "https://github.com/bertrandmbanwi/infraguard"
Issues = "https://github.com/bertrandmbanwi/infraguard/issues"
Changelog = "https://github.com/bertrandmbanwi/infraguard/blob/main/CHANGELOG.md"

[project.scripts]
infraguard = "infraguard.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = [
    "E501",     # line length (handled by formatter)
    "B008",     # function call in defaults (standard Typer pattern)
    "B904",     # raise from in except (noisy for typer.Exit)
    "RUF001",   # ambiguous unicode (intentional emoji)
    "SIM108",   # ternary suggestion (readability preference)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
